hd-contract-agent/requirements.txt
msal>=1.24.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pdfplumber>=0.10.3
python-dotenv>=1.0.0
